+++ src/Chapter3/keysAndBoxes/Item.java
@@ -1,4 +1,4 @@
-package Chapter3;
+package chapter3.keysAndBoxes;
 
 public class Item extends Material {
     private final String name;
