+++ src/Chapter3/keysAndBoxes/Material.java
@@ -1,4 +1,4 @@
-package Chapter3;
+package chapter3.keysAndBoxes;
 
 public abstract class Material {
     abstract public String getName();
