+++ src/Main.java
@@ -1,7 +1,3 @@
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Random;
-
 public class Main {
 
     public static void main(String[] args) {
